requirements.txt
openai==0.28
telegram
python-telegram-bot
python-dotenv